Correction PIB

+++ models/1_staging/stg_olympics__dictionnaire.sql
@@ -7,7 +7,7 @@ renamed AS (
         Pays as lb_pays,
         Code as cd_pays,
         safe_cast(Population as int) as nb_population,
-        safe_cast(Pib_par_habitant as int) as mt_pib_par_habitant
+        safe_cast(Pib_par_habitant as decimal) as mt_pib_par_habitant
     FROM source
 )
 SELECT * from renamed
